+++ client/src/pages/create/create.module.css
@@ -1,0 +1,7 @@
+.container {
+  display: flex;
+  flex-direction: column;
+  width: 100%;
+  height: 100vh;
+  background-color: #e9e5de;
+}
