+++ client/src/pages/index.module.css
@@ -1,7 +1,26 @@
 .container {
+  position: relative;
   display: flex;
   flex-direction: column;
   width: 100%;
   height: 100vh;
   background-color: #e9e5de;
+}
+
+.mainContent {
+  flex: 1;
+  background-image: url('../assets/HomeThinker.jpg');
+  background-repeat: no-repeat;
+  background-position: right top;
+  background-size: no-repeat;
+}
+
+.welcomeText {
+  position: absolute;
+  bottom: 10%;
+  left: 5%;
+  font-family: 'Noto Sans JP', sans-serif;
+  font-size: 3rem;
+  font-weight: bold;
+  color: #000;
 }
